[Refactoring]Define `vram` module
