+++ bootx64/src/mem/paging.rs
@@ -2,6 +2,7 @@
 
 use common_items::constant::*;
 use common_items::size::{Byte, Size};
+use common_items::vram;
 use uefi::table::boot;
 use uefi::table::boot::{AllocateType, MemoryType};
 use x86_64::addr::{PhysAddr, VirtAddr};
@@ -50,7 +51,7 @@ unsafe impl<'a> FrameAllocator<Size4KiB> for AllocatorWithEfiMemoryMap<'a> {
 
 pub fn init(
     boot_services: &boot::BootServices,
-    vram: &common_items::VramInfo,
+    vram: &vram::Info,
     addr_kernel: PhysAddr,
     bytes_kernel: Size<Byte>,
     stack_addr: PhysAddr,
