+++ common_items/src/lib.rs
@@ -1,11 +1,12 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
- 
+
 #![no_std]
 #![feature(const_fn)]
 
 pub mod constant;
 pub mod debug;
 pub mod size;
+pub mod vram;
 
 extern crate uefi;
 extern crate x86_64;
@@ -13,47 +14,7 @@ extern crate x86_64;
 use crate::constant::INIT_RSP;
 use core::ptr;
 use size::{Byte, Size};
-use uefi::proto::console::gop;
 use uefi::table::boot;
-use x86_64::addr::PhysAddr;
-
-#[derive(Copy, Clone)]
-#[repr(C)]
-pub struct VramInfo {
-    bpp: usize,
-    screen_x: usize,
-    screen_y: usize,
-    ptr: PhysAddr,
-}
-
-impl VramInfo {
-    pub fn new_from_gop(gop: &mut gop::GraphicsOutput) -> Self {
-        let (screen_x, screen_y) = gop.current_mode_info().resolution();
-
-        Self {
-            bpp: 32,
-            screen_x: screen_x,
-            screen_y: screen_y,
-            ptr: PhysAddr::new(gop.frame_buffer().as_mut_ptr() as u64),
-        }
-    }
-
-    pub fn bpp(&self) -> usize {
-        self.bpp
-    }
-
-    pub fn resolution(&self) -> (usize, usize) {
-        (self.screen_x, self.screen_y)
-    }
-
-    pub fn phys_ptr(&self) -> PhysAddr {
-        self.ptr
-    }
-
-    pub fn bytes(&self) -> Size<Byte> {
-        Size::new(self.screen_x as usize * self.screen_y as usize * self.bpp as usize / 8)
-    }
-}
 
 #[repr(C)]
 pub struct MemMapInfo {
@@ -75,19 +36,19 @@ impl MemMapInfo {
 
 #[repr(C)]
 pub struct BootInfo {
-    vram_info: VramInfo,
+    vram_info: vram::Info,
     mem_map_info: MemMapInfo,
 }
 
 impl BootInfo {
-    pub fn new(vram_info: VramInfo, mem_map_info: MemMapInfo) -> Self {
+    pub fn new(vram_info: vram::Info, mem_map_info: MemMapInfo) -> Self {
         Self {
             vram_info,
             mem_map_info,
         }
     }
 
-    pub fn vram(&self) -> VramInfo {
+    pub fn vram(&self) -> vram::Info {
         self.vram_info
     }
